third_party/WebKit/LayoutTests/external/wpt/html/browsers/history/the-history-interface/history_go_undefined-expected.txt

This is a testharness.js-based test.
FAIL history.forward() with session history assert_array_equals: Pages opened during history navigation property 2, expected 2 but got 3
Harness: the test ran to completion.
